Refactor: layers directory separated
